requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.11.7

python-multipart==0.0.6
python-dotenv==1.1.0
supabase==2.18.0

httpx>=0.26,<0.29


google-adk
google-generativeai==0.8.5
litellm==1.66.3
python-json-logger==2.0.7
asyncio-throttle==1.0.2
tenacity==8.2.3
structlog==23.2.0

google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5

